requirements.txt
Jinja2==3.1.6
pydantic==2.11.3
PyJWT==2.10.1
requests==2.32.3